refactor(pipeline): extract transformations to src/pipeline/ (refs #298)
